[last] touch

+++ WeatherApplication.Server/DTOs/CurrentWeatherDto.cs
@@ -11,9 +11,9 @@ namespace WeatherApplication.Server.DTOs
     }
     public class Main
     {
-        public int Temp { get; set; }
-        public int Pressure { get; set; }
-        public int Humidity { get; set; }
+        public double Temp { get; set; }
+        public double Pressure { get; set; }
+        public double Humidity { get; set; }
     }
     public class Wind
     {
@@ -21,6 +21,6 @@ namespace WeatherApplication.Server.DTOs
     }
     public class Clouds
     {
-        public int All { get; set; }
+        public double All { get; set; }
     }
 }
